% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/githubstats.R
\name{adds_per_user}
\alias{adds_per_user}
\title{adds_per_user}
\usage{
adds_per_user(current_user)
}
\arguments{
\item{current_user}{current user}
}
\value{
dataframe based on user input
}
\description{
stats of user in terms of additions for the requested repository by the end user.
}
\examples{
\dontrun{
adds_per_user('mattip')
}
}
