% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_jokes.R
\name{all_jokes}
\alias{all_jokes}
\title{Get all possible Christmas joke}
\usage{
all_jokes()
}
\value{
A long list of 2-item-lists containing both the setups (\code{$question}) and punchlines (\code{$answer}).
}
\description{
This function requests all of the jokes from the Christmas Countdown API,
then returns the setups and punchlines in a long list.
}
\examples{
all_jokes()
}
