% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_median_bar_plot.R
\name{average_median_bar_plot}
\alias{average_median_bar_plot}
\title{average median bar plot drawing}
\arguments{
\item{new_dataset}{a data frame}
}
\value{
plots
}
\description{
draw plot of a chosen year and occupation
}
