% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/githubstats.R
\name{commits_per_user_viz}
\alias{commits_per_user_viz}
\title{commits_per_user_viz}
\usage{
commits_per_user_viz(current_user)
}
\arguments{
\item{current_user}{current user}
}
\value{
dataframe based on user input
}
\description{
it gives the option of showing visualization/trends/time series plots of commits of the github user for the requested repository by the end user.
}
\examples{
\dontrun{
commits_per_user_viz('mattip')
}
}
