% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/githubstats.R
\name{deletes_per_user}
\alias{deletes_per_user}
\title{deletes_per_user}
\usage{
deletes_per_user(current_user)
}
\arguments{
\item{current_user}{current user}
}
\value{
dataframe based on user input
}
\description{
stats of user in terms of deletions for the requested repository by the end user.
}
\examples{
\dontrun{
deletes_per_user('mattip')
}
}
