% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/githubstats.R
\name{git_stats}
\alias{git_stats}
\title{git_stats}
\usage{
git_stats(repo_name)
}
\arguments{
\item{repo_name}{A repo}
}
\value{
dataframe and visualization of thr above mentioned based on user input
}
\description{
stats of user in terms of additions, commits, deletions for the requested repository by the end user.
In addition to that it also gives the option of showing visualization/trends/time series plots of additions,commits,
deletions of the github user for the requested repository by the end user.
}
\examples{
\dontrun{
git_stats('numpy/numpy')
}
}
