% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joke.R
\name{joke}
\alias{joke}
\title{Get a Christmas joke}
\usage{
joke()
}
\value{
A list containing both the setup (\code{$question}) and punchline (\code{$answer}).
}
\description{
This function request a random joke from the Christmas Countdown API,
then returns the setup and punchline in a list.
}
\examples{
joke()
}
