% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_occupations.R
\name{list_occupations}
\alias{list_occupations}
\title{List Occupations Matching a Keyword in a Dataset}
\usage{
list_occupations(dataset, str)
}
\arguments{
\item{dataset}{A tibble or data frame containing the wage dataset.}

\item{str}{A character string representing the keyword to match against occupations.}
}
\value{
A character vector containing unique occupations that match the provided keyword. If no matches are found, a message is returned indicating so.
}
\description{
This function takes a wage dataset and a keyword as input and returns a list of unique occupations that match the provided keyword.
The comparison is case-insensitive.
}
\examples{
# List occupations containing the keyword "engineer"
\dontrun{
engineer_occupations <- list_occupations(dataset = your_dataset, str = "engineer")

# List occupations containing the keyword "data"
data_occupations <- list_occupations(dataset = your_dataset, str = "data")

# Handle no matches
no_matches <- list_occupations(dataset = your_dataset, str = "nonexistent")
# This will return a message indicating that there are no matching occupations.
}
}
