% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_NA.R
\name{process_NA}
\alias{process_NA}
\title{identify and process missing values}
\arguments{
\item{dataset}{a data frame}
}
\value{
a dataframe without missing values
}
\description{
Function to identify NAs and draw plots
}
