% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_outlier.R
\name{process_outlier}
\alias{process_outlier}
\title{identify and process outliers}
\arguments{
\item{dataset}{a data frame}
}
\value{
list of occupations that have outliers and outlier visualization
}
\description{
Function to identify outliers and draw plots
}
