% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_left.R
\name{time_left}
\alias{time_left}
\title{Access /timeleft Endpoint}
\usage{
time_left(timezone = NULL)
}
\arguments{
\item{timezone}{character vector, matching a time zone abbreviation. The default value, NULL, retrieves results for UTC.}
}
\value{
a named list containing time until Christmas in months, weeks, sleeps, days, hours, minutes and seconds.
}
\description{
This function gets the time left until Christmas in a variety of different time units.
}
\examples{
time_left()
time_left(timezone = "PST")

}
