% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wage_trend.R
\name{trend}
\alias{trend}
\title{A description of get_wage function.}
\arguments{
\item{records}{The data user put in}

\item{type}{A choice using "overall" or "specific"}

\item{filte}{to remove the outliers}

\item{provs}{a value or vector about province}

\item{positions}{a value or vector about job type}
}
\value{
One or two plot
}
\description{
This is my function to plot the trend.
}
\examples{
trend("overall")
}
