% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weekday.R
\name{weekday}
\alias{weekday}
\title{Access /weekday endpoint}
\usage{
weekday(return_index = FALSE)
}
\arguments{
\item{return_index}{Whether to return the day as a 1-based index integer.
Defaults to FALSE}
}
\value{
A string or an integer corresponding to the weekday that CHristmas
falls on
}
\description{
This function accesses the /weekday endpoint of the Christmas Countdown API.
The API returns an integer corresponding to the weekday that Christmas falls
on. By default, this function returns a string representing the weekday, but
an argument can be passed to specify returning the 1-based index integer
instead.
}
\examples{
weekday()
weekday(return_index = TRUE)

}
\keyword{day}
\keyword{week}
\keyword{weekday}
